pickle: Relax the requirement #thread_starts == #threads_ends
We're dumping stats after a gem5 exit, which is followed by a series
of PerfPage accesses to record thead starts. However, since gem5 exit
is a memory access, and so are the PerfPage accesses, we don't know
which one leaves the write queue first. As a result, there might
be some PerfPage accesses occur before the gem5 exit. So we only
enforce the constraint that the number of thread starts must be
greater or equal the number of thread ends.

Signed-off-by: Hoa Nguyen <hn@hnpl.org>